Update of the first view

+++ Pythagorean Finder/Model/Result.swift
@@ -8,6 +8,6 @@
 import UIKit
 
 struct Result {
-    let value: Double
+    let value: Float
     let color: UIColor
 }
